backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Async Task Queue
celery==5.3.4
redis==5.0.1

# PDF & Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pdfminer.six==20221105
pytesseract==0.3.10
Pillow==10.1.0

# NLP & AI
spacy==3.7.2
sentence-transformers==2.3.1
openai==1.3.7
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.20.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.29.7
aiofiles==23.2.1
httpx==0.25.2  # Required for Textkernel API calls
python-dateutil==2.8.2  # Enhanced date parsing for spaCy parser

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

